# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PropConnect AI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/leads": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "List Leads",
        "operationId": "list_leads_api_leads_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LeadStage"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search client name or city",
              "title": "Q"
            },
            "description": "Search client name or city"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeadCard"
                  },
                  "title": "Response List Leads Api Leads Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "leads"
        ],
        "summary": "Create Lead",
        "operationId": "create_lead_api_leads_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/leads/summary": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "Stage Summary",
        "operationId": "stage_summary_api_leads_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StageCount"
                  },
                  "type": "array",
                  "title": "Response Stage Summary Api Leads Summary Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/leads/{lead_id}": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "Read Lead",
        "operationId": "read_lead_api_leads__lead_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "leads"
        ],
        "summary": "Update Lead",
        "operationId": "update_lead_api_leads__lead_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/leads/{lead_id}/move": {
      "post": {
        "tags": [
          "leads"
        ],
        "summary": "Move Lead",
        "description": "Move a lead to another pipeline stage. Every move is recorded in the stage history.",
        "operationId": "move_lead_api_leads__lead_id__move_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadMove"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/status": {
      "get": {
        "tags": [
          "ai"
        ],
        "summary": "Ai Status",
        "operationId": "ai_status_api_ai_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/leads/{lead_id}/ai/next-step": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Next Step",
        "description": "Ask the AI Lead Assistant for the single most useful next step on this lead.",
        "operationId": "next_step_api_leads__lead_id__ai_next_step_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/leads/{lead_id}/ai/suggestions": {
      "get": {
        "tags": [
          "ai"
        ],
        "summary": "List Suggestions",
        "operationId": "list_suggestions_api_leads__lead_id__ai_suggestions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuggestionRead"
                  },
                  "title": "Response List Suggestions Api Leads  Lead Id  Ai Suggestions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/suggestions/{suggestion_id}/decision": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Decide",
        "description": "The agent accepts or dismisses a suggestion. Accepting logs it on the client's timeline.",
        "operationId": "decide_api_ai_suggestions__suggestion_id__decision_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "suggestion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Suggestion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestionDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AiStatus": {
        "properties": {
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "ai_enabled",
          "model"
        ],
        "title": "AiStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeadCard": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "stage": {
            "$ref": "#/components/schemas/LeadStage"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "client_type": {
            "type": "string",
            "title": "Client Type"
          },
          "budget_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Max"
          },
          "preferred_city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred City"
          },
          "last_contact_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Contact At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stage",
          "source",
          "score",
          "priority",
          "client_id",
          "client_name",
          "client_type",
          "budget_max",
          "preferred_city",
          "last_contact_at",
          "updated_at"
        ],
        "title": "LeadCard"
      },
      "LeadCreate": {
        "properties": {
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "source": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "client_id"
        ],
        "title": "LeadCreate"
      },
      "LeadDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "stage": {
            "$ref": "#/components/schemas/LeadStage"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "lost_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lost Reason"
          },
          "agent_id": {
            "type": "integer",
            "title": "Agent Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "client": {
            "$ref": "#/components/schemas/ClientRead"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/StageHistoryRead"
            },
            "type": "array",
            "title": "History"
          },
          "score_breakdown": {
            "items": {
              "$ref": "#/components/schemas/ScoreItemRead"
            },
            "type": "array",
            "title": "Score Breakdown"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stage",
          "source",
          "score",
          "priority",
          "lost_reason",
          "agent_id",
          "created_at",
          "updated_at",
          "client",
          "history",
          "score_breakdown"
        ],
        "title": "LeadDetail"
      },
      "LeadMove": {
        "properties": {
          "to_stage": {
            "$ref": "#/components/schemas/LeadStage"
          },
          "lost_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Lost Reason"
          }
        },
        "type": "object",
        "required": [
          "to_stage"
        ],
        "title": "LeadMove"
      },
      "LeadStage": {
        "type": "string",
        "enum": [
          "new",
          "contacted",
          "qualified",
          "proposal",
          "negotiation",
          "won",
          "lost"
        ],
        "title": "LeadStage"
      },
      "LeadUpdate": {
        "properties": {
          "source": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "title": "LeadUpdate"
      },
...
      "ScoreItemRead": {
        "properties": {
          "factor": {
            "type": "string",
            "title": "Factor"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "max_points": {
            "type": "integer",
            "title": "Max Points"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "factor",
          "points",
          "max_points",
          "reason"
        ],
        "title": "ScoreItemRead"
      },
      "StageCount": {
        "properties": {
          "stage": {
            "$ref": "#/components/schemas/LeadStage"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "stage",
          "count"
        ],
        "title": "StageCount"
      },
      "StageHistoryRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "from_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LeadStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_stage": {
            "$ref": "#/components/schemas/LeadStage"
          },
          "changed_by_id": {
            "type": "integer",
            "title": "Changed By Id"
          },
          "changed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Changed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "from_stage",
          "to_stage",
          "changed_by_id",
          "changed_at"
        ],
        "title": "StageHistoryRead"
      },
      "SuggestionDecision": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/SuggestionStatus",
            "description": "accepted or dismissed"
          },
          "draft_message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1500
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Message",
            "description": "The agent's edited draft, if changed"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SuggestionDecision"
      },
      "SuggestionRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lead_id": {
            "type": "integer",
            "title": "Lead Id"
          },
          "client_id": {
            "type": "integer",
            "title": "Client Id"
          },
          "source": {
            "$ref": "#/components/schemas/SuggestionSource"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "intent_level": {
            "type": "string",
            "title": "Intent Level"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "recommended_action": {
            "type": "string",
            "title": "Recommended Action"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "reasoning": {
            "type": "string",
            "title": "Reasoning"
          },
          "draft_message": {
            "type": "string",
            "title": "Draft Message"
          },
          "status": {
            "$ref": "#/components/schemas/SuggestionStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lead_id",
          "client_id",
          "source",
          "model",
          "intent_level",
          "summary",
          "recommended_action",
          "action_type",
          "reasoning",
          "draft_message",
          "status",
          "created_at",
          "decided_at"
        ],
        "title": "SuggestionRead"
      },
      "SuggestionSource": {
        "type": "string",
        "enum": [
          "openai",
          "rules"
        ],
        "title": "SuggestionSource"
      },
      "SuggestionStatus": {
        "type": "string",
        "enum": [
          "suggested",
          "accepted",
          "dismissed"
        ],
        "title": "SuggestionStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}